requirements.txt
click~=8.1.3
pymongo~=4.3.3
requests~=2.28.2
beautifulsoup4~=4.11.1
selenium~=4.6.0
python-dotenv~=0.19.2
python-cli-scheduler